backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
ultralytics==8.0.206
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3
pydantic==2.5.0